feat(olympus): add /why and /system destinations; redirect legacy routes
- /why hosts the research client; /system hosts the observability dashboard
  (internal reorg deferred to Phases 4 & 5)
- /research -> /why, /observability -> /system, /library -> /why (param-preserving)
- mobile nav is the sidebar drawer (no separate destination bar); Task 1.3 dropped
